requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=1.5.0
python-dateutil>=2.8.0
typing-extensions>=4.0.0
